SAK-52658 SiteStats wire report access helper
