+++ sitestats/sitestats-impl/src/java/org/sakaiproject/sitestats/impl/view/SiteStatsViewServiceImpl.java
@@ -12,50 +12,25 @@ import lombok.Setter;
 
 import org.apache.commons.lang3.StringUtils;
 import org.sakaiproject.sitestats.api.PrefsData;
-import org.sakaiproject.sitestats.api.StatsAuthz;
 import org.sakaiproject.sitestats.api.StatsManager;
 import org.sakaiproject.sitestats.api.report.Report;
 import org.sakaiproject.sitestats.api.report.ReportDef;
 import org.sakaiproject.sitestats.api.report.ReportManager;
 import org.sakaiproject.sitestats.api.view.SiteStatsOverview;
-import org.sakaiproject.sitestats.api.view.SiteStatsReportPreviewService;
 import org.sakaiproject.sitestats.api.view.SiteStatsReportRequest;
 import org.sakaiproject.sitestats.api.view.SiteStatsReportSummary;
 import org.sakaiproject.sitestats.api.view.SiteStatsReportView;
 import org.sakaiproject.sitestats.api.view.SiteStatsViewService;
 import org.sakaiproject.sitestats.api.view.SiteStatsWidgetMetric;
 import org.sakaiproject.sitestats.api.view.SiteStatsWidgetTab;
-import org.sakaiproject.tool.api.SessionManager;
 
 public class SiteStatsViewServiceImpl implements SiteStatsViewService {
 
 	@Setter private StatsManager statsManager;
 	@Setter private SiteStatsReportViewMapper siteStatsReportViewMapper;
-	private ReportManager reportManager;
-	private SiteStatsWidgetCatalog siteStatsWidgetCatalog;
-	private SiteStatsReportAccess siteStatsReportAccess = new SiteStatsReportAccess();
-
-	public void setStatsAuthz(StatsAuthz statsAuthz) {
-		siteStatsReportAccess.setStatsAuthz(statsAuthz);
-	}
-
-	public void setReportManager(ReportManager reportManager) {
-		this.reportManager = reportManager;
-		siteStatsReportAccess.setReportManager(reportManager);
-	}
-
-	public void setSiteStatsReportPreviewService(SiteStatsReportPreviewService siteStatsReportPreviewService) {
-		siteStatsReportAccess.setSiteStatsReportPreviewService(siteStatsReportPreviewService);
-	}
-
-	public void setSessionManager(SessionManager sessionManager) {
-		siteStatsReportAccess.setSessionManager(sessionManager);
-	}
-
-	public void setSiteStatsWidgetCatalog(SiteStatsWidgetCatalog siteStatsWidgetCatalog) {
-		this.siteStatsWidgetCatalog = siteStatsWidgetCatalog;
-		siteStatsReportAccess.setSiteStatsWidgetCatalog(siteStatsWidgetCatalog);
-	}
+	@Setter private ReportManager reportManager;
+	@Setter private SiteStatsWidgetCatalog siteStatsWidgetCatalog;
+	@Setter private SiteStatsReportAccess siteStatsReportAccess;
 
 	@Override
 	public SiteStatsOverview getOverview(String siteId) {
